Create core directory, move system to its own file
